Updated the source files to reflect sns_ik and lib move
