+++ sns_ik_lib/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package>
-  <name>sns_ikl</name>
+  <name>sns_ik_lib</name>
   <version>0.1.0</version>
   <description>
    Saturation in the Null Space (SNS) Inverse Kinematic Library.
